refactor(evals): colocate base classes with implementations
