+++ dimos/evals/agents/lib/single_call.py
@@ -26,7 +26,7 @@ from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, System
 from langchain_core.outputs import ChatGeneration
 
 from dimos.agents.llm_trace import latest_pair, write_normalized
-from dimos.evals.agent import Agent, AgentConfig
+from dimos.evals.agents.base import Agent, AgentConfig
 from dimos.evals.agents.lib.langchain_to_atif import append_ai_message_to_atif
 from dimos.evals.agents.lib.trajectory_builder import TrajectoryBuilder
 from dimos.evals.types import RunningEnvironment, Trajectory
