+++ dimos/evals/environments/dataset.py
@@ -21,7 +21,7 @@ from typing import TYPE_CHECKING, Any
 
 from dimos.agents.mcp.mcp_adapter import McpAdapter
 from dimos.e2e_tests.dimos_cli_call import DimosCliCall
-from dimos.evals.environment import Environment
+from dimos.evals.environments.base import Environment
 from dimos.evals.environments.lib.launch import default_mcp_url, validate_blueprints
 from dimos.evals.types import RunningEnvironment
 from dimos.memory.cli.dataset import open_dataset, resolve_dataset
@@ -30,7 +30,7 @@ from dimos.memory.stream import Stream
 from dimos.protocol.service.spec import BaseConfig
 
 if TYPE_CHECKING:
-    from dimos.evals.agent import Agent
+    from dimos.evals.agents.base import Agent
 
 
 class DatasetConfig(BaseConfig):
