+++ dimos/evals/environments/sim.py
@@ -28,13 +28,13 @@ from dimos.constants import RECORDINGS_DIR
 from dimos.core.run_registry import list_runs
 from dimos.e2e_tests.dim_sim_client import DimSimClient
 from dimos.e2e_tests.dimos_cli_call import DimosCliCall
-from dimos.evals.environment import Environment
+from dimos.evals.environments.base import Environment
 from dimos.evals.environments.lib.launch import default_mcp_url, validate_blueprints
 from dimos.evals.types import RunningEnvironment
 from dimos.protocol.service.spec import BaseConfig
 
 if TYPE_CHECKING:
-    from dimos.evals.agent import Agent
+    from dimos.evals.agents.base import Agent
     from dimos.memory.store.base import Store
